Make CoreDataModel conform ModelConvertible
